% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocFileCache.R
\name{blast_db_cache}
\alias{blast_db_cache}
\alias{blast_db_get}
\title{Manage BLAST Database Downloads using BioCFileCache}
\usage{
blast_db_cache()

blast_db_get(
  file = "16S_ribosomal_RNA.tar.gz",
  baseURL = "https://ftp.ncbi.nlm.nih.gov/blast/db/",
  verbose = TRUE
)
}
\arguments{
\item{file}{the filename of the database.}

\item{baseURL}{URL to download blast databases from. The default is NCBI's
ftp server.}

\item{verbose}{logical; display download information.}
}
\value{
\itemize{
\item \code{blast_db_cache()} returns the path to the local \link{BiocFileCache} cache.
\item \code{blast_db_get()} returns the file path to a downloaded BLAST database
file.
}
}
\description{
Use \link{BiocFileCache} to manage local copies of BLAST database downloads.
NCBI BLAST databases are updated daily and
may be downloaded via FTP from \url{https://ftp.ncbi.nlm.nih.gov/blast/db/}.
}
\examples{
library(BiocFileCache)

## show the package's cache directory
cache <- blast_db_cache()
cache

## get a database file (will be downloaded the first time)
db_16S <- blast_db_get("16S_ribosomal_RNA.tar.gz")
db_16S

## bfc functions can be used to manage the local cache
bfcinfo(cache)
}
\seealso{
Other blast: 
\code{\link{blast}()},
\code{\link{makeblastdb}()}
}
\author{
Michael Hahsler
}
\concept{blast}
