% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA_rBLAST-package.R
\docType{package}
\name{rBLAST-package}
\alias{rBLAST}
\alias{rBLAST-package}
\title{rBLAST: R Interface for the Basic Local Alignment Search Tool}
\description{
Seamlessly interfaces the Basic Local Alignment Search Tool (BLAST) to search genetic sequence data bases. This work was partially supported by grant no. R21HG005912 from the National Human Genome Research Institute.
}
\section{Citation Request}{


Cite the use of this package as:

Hahsler M, Nagar A (2019). rBLAST: R Interface for the Basic Local
Alignment Search Tool. R package version 0.99.2,
URL: https://github.com/mhahsler/rBLAST.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mhahsler/rBLAST}
  \item Report bugs at \url{https://github.com/mhahsler/rBLAST/issues}
}

}
\author{
\strong{Maintainer}: Michael Hahsler \email{mhahsler@lyle.smu.edu} (\href{https://orcid.org/0000-0003-2716-1405}{ORCID})

Authors:
\itemize{
  \item Nagar Anurag
}

}
\keyword{internal}
